Update to image similarity stuff.

+++ src/python/turicreate/toolkits/_image_feature_extractor.py
@@ -181,7 +181,7 @@ class TensorFlowFeatureExtractor(ImageFeatureExtractor):
             batch_store = state['batch_store']
             
             if not batch_store:
-                batch_store.append(np.zeros((0,)))  # Will get resized later
+                batch_store.append(np.zeros((0,), dtype=np.float32))  # Will get resized later
 
             return batch_store.pop()
